The update for the light information is working
Its still not working as easily as i would like because the user still has to hit save but its working nonetheless

The problem was that because I initially made the keys customizable, it doesnt seem that firebase allows you to easily modify the key name, also since the position in the recycler view which the user clicked shows is important, if we delete the old node and create a new one the information cant populate because the app no longer knows which light information it wants, the position wont stay the same since firebase is in alphabetical order.

To solve that issue what i did was made the keys for each light static, they are light 1, light 2 and light 3 and then when the user updates the light information it doesnt need to grab a new postiion or anything because itll still be the same child node. only the sub fields will need to get updated.